LPD-82339 Move tests to existing corresponding file
